Tidy up OWL rendering code.
The rendering objects are no longer keeping OWLAxiom objects as fields.
The axiom construction should happen at the higher code level.

The ReferenceRendering is redundant. The rendering acts like an
"intermediate state" from a cell reference to a value with a type and
then to an OWL rendering (e.g., @A1 --> "Person" as OWLClass -->
OWLClass(Person)). We can just cut it short from a cell reference to an
OWL rendering (e.g., @A1 --> OWLClass(Person)).

We are deleting those reference rendering classes.