Simplify directive settings module.
Remove the 'settings' package. This package contains all the enums for
mapping-master directives. However, all the enumeration values come from
MappingMasterParserConstants which is generated by javacc, and the enums
are just a wrapper for these values.

As a result, this approach will give a hard time for maintenance without
giving further value. The directive grouping can be observed in the
grammar and client code can just use MappingMasterParserConstants
directly.

I keep using the ReferenceDirectives and move it to 'directive' package.