Deprecate rendering module.
Remove the 'rendering' package from the code base.

The rendering module provides the abstraction to store temporarily the
rendering objects and return the end result collectively (e.g., a set of
axioms or a complete text).

These intermediate 'containers' are proven to be redundant and hard to
maintain. Because each renderer's returned object is required to have a
dedicated rendering class before the renderer produces the real output
objects.

In addition, since we won't rely anymore to the old renderer, this
package will become obsolete and won't be used anymore. Deleting them is
necessary.