Simplify the public API to instantiate mapping-master
No more intermediate/wrapper classes for ontology and workbook. The
application framework is using OWLAPI and Apache POI objects
immediately.

It may reduce the extensibility of the API but surely it will be easier
to maintain. Unless there is a foreseeable plan to use different
ontology languages or data sources, this simplification is favorable.