Remove javacc plugin from Maven
It turns out to be confusing to have the grammar compilation as part of
Maven install phase. So I removed it.

If you are updating the grammar then you need to compile it through the
command line and do the manual file transfer. Documentation and script
will be provided.

Executing `mvn clean install` will just compile all source code in
"scr/main/java" directory only, without making it complicated with
recompiling the grammar and files at "target/generated-sources/javacc".