Fix the grammar to handle data and object property ambiguity
Solve the ambiguity at the higher level. Use the renderer's entity
resolver to decide if the given property name is data property or object
property, instead of having the decision when parsing the name (i.e., at
the lower level).

As a result, the grammar will just produce an untyped property and
untyped property value, and let the entity resolver decides the exact
type later.

In case of having a reference notation (i.e., '@...'), users still have
to give a hint about the property type, or otherwise the renderer will
apply the default type.