Add test cases for the if-entity-absent directives.
Note that there are 3 tests that are ignored currently, i.e.,
shouldIgnoreIfEntityIsAbsent(), shouldWarnIfEntityIsAbsent() and
shouldFailIfEntityIsAbsent(). The reason is because the API is no longer
interact with the reference directives once it reaches the OWL object
creation code.

The fix will require a small refactor to allow the OwlFactory to access
the reference directives. However, since the usage of these directives
is an edge case, we are going to put this issue as a pending case.